+++ src/components/pbdetailPage/Intro.tsx
@@ -1,78 +1,40 @@
 "use client";
-import React, { useState } from "react";
+import React from "react";
 import Link from "next/link";
 import Image from "next/image";
 import { useRoleStore } from "@/store/roleStore";
 import BlurModal from "@/components/common/Modal/BlurModal";
 import ButtonModal from "@/components/common/ButtonModal";
 import { usePathname, useRouter } from "next/navigation";
-// import shareKakao from "@/utils/shareKakao";
+import useBookMark from "@/hooks/useBookMark";
+import useShare from "@/hooks/useShare";
 
 function Intro({ introData, edit }: { introData: any; edit: boolean }) {
   const { profile, name, isBookmarked, branchName, msg, companyId,companyName, companyLogo, reserveCount, reviewCount } =
     introData;
-  const [isBookmark, setIsBookmark] = useState(isBookmarked);
-  const [isBookmarkOpen, setIsBookmarkOpen] = useState(false);
-  const [isShare, setIsShare] = useState(false);
-  const [isShareOpen, setIsShareOpen] = useState(false);
-  const [isCopy, setIsCopy] = useState(false);
-  const [isCopyOpen, setIsCopyOpen] = useState(false);
   const { getRole } = useRoleStore();
   const pathname = usePathname();
   const router = useRouter();
   const base = "http://localhost:3000";
   const urlToCopy = base + pathname;
+  const { isBookmark, isBookmarkOpen, setIsBookmarkOpen, bookMarkHandler, bookMarkContents } = useBookMark(
+    isBookmarked,
+    "/bookmark/pb",
+  );
+  const {
+    isShare,
+    isShareOpen,
+    setIsShareOpen,
+    shareHandler,
+    shareContents,
+    isCopy,
+    isCopyOpen,
+    setIsCopyOpen,
+    copyContents,
+  } = useShare(urlToCopy);
 
   const goToCompany = () => {
     router.push(`/pblist/financial/${companyId}`);
-  };
-
-  const bookMarkHandler = () => {
-    setIsBookmarkOpen(true);
-    setIsBookmark(!isBookmark);
-    // 북마크 여부에 따라 추가, 삭제 api호출
-  };
-
-  const shareHandler = () => {
-    setIsShareOpen(true);
-    setIsShare(!isShare);
-  };
-
-  const shareContents = {
-    content: "PB 정보 공유하기",
-    confirmText: "카카오톡으로 공유",
-    cancelText: "링크 복사",
-    confirmFn: () => {
-      // shareKakao(base, "테스트", "https://www.google.com", "/public/assets/images/default_profile.png");
-      setIsShareOpen(false);
-    },
-    cancelFn: () => {
-      navigator.clipboard.writeText(urlToCopy);
-      setIsShareOpen(false);
-      setIsCopy(!isCopy);
-      setIsCopyOpen(true);
-    },
-  };
-
-  const copyContents = {
-    content: "링크가 복사되었습니다.",
-    confirmText: "확인",
-    confirmFn: () => {
-      setIsCopyOpen(false);
-    },
-  };
-
-  const bookMarkContents = {
-    content: "북마크에 추가되었습니다.",
-    confirmText: "확인",
-    cancelText: "북마크 바로가기",
-    confirmFn: () => {
-      setIsBookmarkOpen(false);
-    },
-    cancelFn: () => {
-      router.push("/bookmark/pb");
-      setIsBookmarkOpen(false);
-    },
   };
 
   return (
